+++ base/resetpasswd.php
@@ -35,7 +35,7 @@ $language = $app->language();
 \bindtextdomain('messages', '../locale');
 \textdomain('messages');
 // Load the menu and assign it to a SMARTY Variable
-$mainmenu = $app->config()->readMenu('mainmenu');
+$mainmenu = $app->menus()->readMenu('mainmenu');
 $app->tpl()->assign('MAINMENU', $mainmenu);
 
 /* Send the HTTP Headers required by the application */